refactor: refactored login as guest
